pyproject.toml
[project]
name = "kaggle-aimo2-fast-math-r1"
version = "0.1.0"
description = ""
authors = [
    {name = "Hiroshi Yoshihara"}
]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
kaggle = ">=1.6.17,<2.0.0"
torch = {version = "^2.5.0", source = "torch_cu121"}
torchvision = {version = "^0.20.0", source = "torch_cu121"}
torchaudio = {version = "^2.5.0", source = "torch_cu121"}
transformers = "^4.47.0"
accelerate = "^1.3.0"
bitsandbytes = "^0.45.0"
sentencepiece = "^0.2.0"
tiktoken = "^0.7.0"
vllm = "0.7.3"
numpy = "1.26.4"
pandas = "^2.2.3"
setuptools = "^75.8.0"
ipykernel = "^6.29.5"
datasets = "^3.2.0"
trl = "0.15.2"
peft = "^0.14.0"
wandb = "^0.19.6"
deepspeed = "0.16.4"
logits-processor-zoo = "^0.1.3"
python-levenshtein = "^0.26.1"
word2number = "^1.1"
flashinfer-python = "0.2.2.post1"
math-verify = "^0.7.0"
kagglehub = "^0.3.10"
seaborn = "^0.13.2"

[[tool.poetry.source]]
name = "torch_cu121"
url = "https://download.pytorch.org/whl/cu121"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
